Implement `octet_counting` framer
