+++ src/internal_events/statsd_source.rs
@@ -1,4 +1,5 @@
 use super::InternalEvent;
+use bytes::Bytes;
 use metrics::counter;
 
 #[derive(Debug)]
@@ -18,19 +19,19 @@ impl InternalEvent for StatsdEventReceived {
 }
 
 #[derive(Debug)]
-pub struct StatsdInvalidRecord<'a> {
+pub struct StatsdInvalidRecord {
     pub error: crate::sources::statsd::parser::ParseError,
-    pub text: &'a str,
+    pub bytes: Bytes,
 }
 
-impl InternalEvent for StatsdInvalidRecord<'_> {
+impl InternalEvent for StatsdInvalidRecord {
     fn emit_logs(&self) {
-        error!(message = "Invalid packet from statsd, discarding.", error = ?self.error, text = %self.text);
+        error!(message = "Invalid packet from statsd, discarding.", error = ?self.error, bytes = %String::from_utf8_lossy(&self.bytes));
     }
 
     fn emit_metrics(&self) {
         counter!("invalid_record_total", 1,);
-        counter!("invalid_record_bytes_total", self.text.len() as u64,);
+        counter!("invalid_record_bytes_total", self.bytes.len() as u64);
     }
 }
