+++ src/sources/syslog.rs
@@ -1,8 +1,8 @@
 use super::util::{SocketListenAddr, TcpSource};
 #[cfg(unix)]
 use crate::sources::util::build_unix_stream_source;
-use crate::udp;
 use crate::{
+    codecs::{BoxedFramingError, OctetCountingCodec},
     config::{
         log_schema, DataType, GenerateConfig, Resource, SourceConfig, SourceContext,
         SourceDescription,
@@ -12,22 +12,18 @@ use crate::{
     shutdown::ShutdownSignal,
     tcp::TcpKeepaliveConfig,
     tls::{MaybeTlsSettings, TlsConfig},
-    Pipeline,
+    udp, Pipeline,
 };
-use bytes::{Buf, Bytes, BytesMut};
+use bytes::Bytes;
 use chrono::{Datelike, Utc};
 use futures::{SinkExt, StreamExt};
 use serde::{Deserialize, Serialize};
-use std::io;
 use std::net::SocketAddr;
 #[cfg(unix)]
 use std::path::PathBuf;
 use syslog_loose::{IncompleteDate, Message, ProcId, Protocol};
 use tokio::net::UdpSocket;
-use tokio_util::{
-    codec::{BytesCodec, Decoder, LinesCodec, LinesCodecError},
-    udp::UdpFramed,
-};
+use tokio_util::{codec::BytesCodec, udp::UdpFramed};
 
 #[derive(Deserialize, Serialize, Debug)]
 // TODO: add back when serde-rs/serde#1358 is addressed
@@ -138,11 +134,15 @@ impl SourceConfig for SyslogConfig {
             #[cfg(unix)]
             Mode::Unix { path } => Ok(build_unix_stream_source(
                 path,
-                SyslogDecoder::new(self.max_length),
+                OctetCountingCodec::new_with_max_length(self.max_length),
                 host_key,
                 cx.shutdown,
                 cx.out,
-                |host_key, default_host, line| Some(event_from_str(host_key, default_host, line)),
+                |host_key, default_host, bytes| {
+                    // Guaranteed to be valid UTF-8 by `OctetCountingCodec`.
+                    let line = std::str::from_utf8(&bytes).unwrap();
+                    Some(event_from_str(host_key, default_host, line))
+                },
             )),
         }
     }
@@ -172,223 +172,17 @@ struct SyslogTcpSource {
 }
 
 impl TcpSource for SyslogTcpSource {
-    type Error = LinesCodecError;
-    type Decoder = SyslogDecoder;
+    type Error = BoxedFramingError;
+    type Decoder = OctetCountingCodec;
 
     fn decoder(&self) -> Self::Decoder {
-        SyslogDecoder::new(self.max_length)
+        OctetCountingCodec::new_with_max_length(self.max_length)
     }
 
-    fn build_event(&self, frame: String, host: Bytes) -> Option<Event> {
-        Some(event_from_str(&self.host_key, Some(host), &frame))
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-enum State {
-    NotDiscarding,
-    Discarding(usize),
-    DiscardingToEol,
-}
-
-/// Decodes according to `Octet Counting` in https://tools.ietf.org/html/rfc6587
-#[derive(Clone, Debug)]
-struct SyslogDecoder {
-    other: LinesCodec,
-    octet_decoding: Option<State>,
-}
-
-impl SyslogDecoder {
-    fn new(max_length: usize) -> Self {
-        Self {
-            other: LinesCodec::new_with_max_length(max_length),
-            octet_decoding: None,
-        }
-    }
-
-    fn octet_decode(
-        &mut self,
-        state: State,
-        src: &mut BytesMut,
-    ) -> Result<Option<String>, LinesCodecError> {
-        // Encoding scheme:
-        //
-        // len ' ' data
-        // |    |  | len number of bytes that contain syslog message
-        // |    |
-        // |    | Separating whitespace
-        // |
-        // | ASCII decimal number of unknown length
-
-        let space_pos = src.iter().position(|&b| b == b' ');
-
-        // If we are discarding, discard to the next newline.
-        let newline_pos = src.iter().position(|&b| b == b'\n');
-
-        match (state, newline_pos, space_pos) {
-            (State::Discarding(chars), _, _) if src.len() >= chars => {
-                // We have a certain number of chars to discard.
-                // There are enough chars in this frame to discard
-                src.advance(chars);
-                self.octet_decoding = None;
-                Err(LinesCodecError::Io(io::Error::new(
-                    io::ErrorKind::Other,
-                    "Frame length limit exceeded",
-                )))
-            }
-
-            (State::Discarding(chars), _, _) => {
-                // We have a certain number of chars to discard.
-                // There aren't enough in this frame so we need to discard
-                // The entire frame and adjust the amount to discard accordingly.
-                self.octet_decoding = Some(State::Discarding(src.len() - chars));
-                src.advance(src.len());
-                Ok(None)
-            }
-
-            (State::DiscardingToEol, Some(offset), _) => {
-                // When discarding we keep discarding to the next newline.
-                src.advance(offset + 1);
-                self.octet_decoding = None;
-                Err(LinesCodecError::Io(io::Error::new(
-                    io::ErrorKind::Other,
-                    "Frame length limit exceeded",
-                )))
-            }
-
-            (State::DiscardingToEol, None, _) => {
-                // There is no newline in this frame. Since we don't have a set number of
-                // chars we want to discard, we need to discard to the next newline.
-                // Advance as far as we can to discard the entire frame.
-                src.advance(src.len());
-                Ok(None)
-            }
-
-            (State::NotDiscarding, _, Some(space_pos)) if space_pos < self.other.max_length() => {
-                // Everything looks good. We aren't discarding, we have a space that is not beyond our
-                // maximum length. Attempt to parse the bytes as a number which will hopefully
-                // give us a sensible length for our message.
-                let len: usize = match std::str::from_utf8(&src[..space_pos])
-                    .map_err(|_| ())
-                    .and_then(|num| num.parse().map_err(|_| ()))
-                {
-                    Ok(len) => len,
-                    Err(_) => {
-                        // It was not a sensible number.
-                        // Advance the buffer past the erroneous bytes
-                        // to prevent us getting stuck in an infinite loop.
-                        src.advance(space_pos + 1);
-                        self.octet_decoding = None;
-                        return Err(LinesCodecError::Io(io::Error::new(
-                            io::ErrorKind::InvalidData,
-                            "Unable to decode message len as number",
-                        )));
-                    }
-                };
-
-                let from = space_pos + 1;
-                let to = from + len;
-
-                if len > self.other.max_length() {
-                    // The length is greater than we want.
-                    // We need to discard the entire message.
-                    self.octet_decoding = Some(State::Discarding(len));
-                    src.advance(space_pos + 1);
-
-                    Ok(None)
-                } else if let Some(msg) = src.get(from..to) {
-                    let s = match std::str::from_utf8(msg) {
-                        Ok(s) => s.to_string(),
-                        Err(_) => {
-                            // The data was not valid UTF8 :-(.
-                            // Advance the buffer past the erroneous bytes
-                            // to prevent us getting stuck in an infinite loop.
-                            src.advance(to);
-                            self.octet_decoding = None;
-                            return Err(LinesCodecError::Io(io::Error::new(
-                                io::ErrorKind::InvalidData,
-                                "Unable to decode message as UTF8",
-                            )));
-                        }
-                    };
-
-                    // We have managed to read the entire message as valid UTF8!
-                    src.advance(to);
-                    self.octet_decoding = None;
-                    Ok(Some(s))
-                } else {
-                    // We have an acceptable number of bytes in this message, but all the data
-                    // was not in the frame, return None to indicate we want more data before we
-                    // do anything else.
-                    Ok(None)
-                }
-            }
-
-            (State::NotDiscarding, Some(newline_pos), _) => {
-                // Beyond maximum length, advance to the newline.
-                src.advance(newline_pos + 1);
-                Err(LinesCodecError::Io(io::Error::new(
-                    io::ErrorKind::Other,
-                    "Frame length limit exceeded",
-                )))
-            }
-
-            (State::NotDiscarding, None, _) if src.len() < self.other.max_length() => {
-                // We aren't discarding, but there is no useful character to tell us what to do next,
-                // we are still not beyond the max length, so just return None to indicate we need to
-                // wait for more data.
-                Ok(None)
-            }
-
-            (State::NotDiscarding, None, _) => {
-                // There is no newline in this frame and we have more data than we want to handle.
-                // Advance as far as we can to discard the entire frame.
-                self.octet_decoding = Some(State::DiscardingToEol);
-                src.advance(src.len());
-                Ok(None)
-            }
-        }
-    }
-
-    /// None if this is not octet counting encoded
-    fn checked_decode(
-        &mut self,
-        src: &mut BytesMut,
-    ) -> Option<Result<Option<String>, LinesCodecError>> {
-        if let Some(&first_byte) = src.get(0) {
-            if (49..=57).contains(&first_byte) {
-                // First character is non zero number so we can assume that
-                // octet count framing is used.
-                trace!("Octet counting encoded event detected.");
-                self.octet_decoding = Some(State::NotDiscarding);
-            }
-        }
-
-        self.octet_decoding
-            .map(|state| self.octet_decode(state, src))
-    }
-}
-
-impl Decoder for SyslogDecoder {
-    type Item = String;
-    type Error = LinesCodecError;
-
-    fn decode(&mut self, src: &mut BytesMut) -> Result<Option<Self::Item>, Self::Error> {
-        if let Some(ret) = self.checked_decode(src) {
-            ret
-        } else {
-            // Octet counting isn't used so fallback to newline codec.
-            self.other.decode(src)
-        }
-    }
-
-    fn decode_eof(&mut self, buf: &mut BytesMut) -> Result<Option<Self::Item>, Self::Error> {
-        if let Some(ret) = self.checked_decode(buf) {
-            ret
-        } else {
-            // Octet counting isn't used so fallback to newline codec.
-            self.other.decode_eof(buf)
-        }
+    fn build_event(&self, frame: Bytes, host: Bytes) -> Option<Event> {
+        // Guaranteed to be valid UTF-8 by `OctetCountingCodec`.
+        let line = std::str::from_utf8(&frame).unwrap();
+        Some(event_from_str(&self.host_key, Some(host), line))
     }
 }
 
@@ -549,7 +343,6 @@ fn insert_fields_from_syslog(event: &mut Event, parsed: Message<&str>) {
 mod test {
     use super::*;
     use crate::{config::log_schema, event::Event};
-    use bytes::BufMut;
     use chrono::prelude::*;
     use shared::assert_event_data_eq;
 
@@ -903,125 +696,5 @@ mod test {
         }
 
         assert_event_data_eq!(event_from_str(&"host".to_string(), None, &raw), expected);
-    }
-
-    #[test]
-    fn non_octet_decode_works_with_multiple_frames() {
-        let mut decoder = SyslogDecoder::new(128);
-        let mut buffer = BytesMut::with_capacity(16);
-
-        buffer.put(&b"<57>Mar 25 21:47:46 gleichner6005 quaerat[2444]: There were "[..]);
-        let result = decoder.decode(&mut buffer);
-        assert_eq!(Ok(None), result.map_err(|_| true));
-
-        buffer.put(&b"8 penguins in the shop.\n"[..]);
-        let result = decoder.decode(&mut buffer);
-        assert_eq!(
-            Ok(Some("<57>Mar 25 21:47:46 gleichner6005 quaerat[2444]: There were 8 penguins in the shop.".to_string())),
-            result.map_err(|_| true)
-        );
-    }
-
-    #[test]
-    fn octet_decode_works_with_multiple_frames() {
-        let mut decoder = SyslogDecoder::new(30);
-        let mut buffer = BytesMut::with_capacity(16);
-
-        buffer.put(&b"28 abcdefghijklm"[..]);
-        let result = decoder.decode(&mut buffer);
-        assert_eq!(Ok(None), result.map_err(|_| false));
-
-        // Sending another frame starting with a number should not cause it to
-        // try to decode a new message.
-        buffer.put(&b"3 nopqrstuvwxyz"[..]);
-        let result = decoder.decode(&mut buffer);
-        assert_eq!(
-            Ok(Some("abcdefghijklm3 nopqrstuvwxyz".to_string())),
-            result.map_err(|_| false)
-        );
-    }
-
-    #[test]
-    fn octet_decode_moves_past_invalid_length() {
-        let mut decoder = SyslogDecoder::new(16);
-        let mut buffer = BytesMut::with_capacity(16);
-
-        // An invalid syslog message that starts with a digit so we think it is starting with the len.
-        buffer.put(&b"232>1 zork"[..]);
-        let result = decoder.decode(&mut buffer);
-
-        assert!(result.is_err());
-        assert_eq!(b"zork"[..], buffer);
-    }
-
-    #[test]
-    fn octet_decode_moves_past_invalid_utf8() {
-        let mut decoder = SyslogDecoder::new(16);
-        let mut buffer = BytesMut::with_capacity(16);
-
-        // An invalid syslog message containing invalid utf8 bytes.
-        buffer.put(&[b'4', b' ', 0xf0, 0x28, 0x8c, 0xbc][..]);
-        let result = decoder.decode(&mut buffer);
-
-        assert!(result.is_err());
-        assert_eq!(b""[..], buffer);
-    }
-
-    #[test]
-    fn octet_decode_moves_past_exceeded_frame_length() {
-        let mut decoder = SyslogDecoder::new(16);
-        let mut buffer = BytesMut::with_capacity(32);
-
-        buffer.put(&b"32thisshouldbelongerthanthmaxframeasizewhichmeansthesyslogparserwillnotbeabletodecodeit\n"[..]);
-        let result = decoder.decode(&mut buffer);
-
-        assert!(result.is_err());
-        assert_eq!(b""[..], buffer);
-    }
-
-    #[test]
-    fn octet_decode_rejects_exceeded_frame_length() {
-        let mut decoder = SyslogDecoder::new(16);
-        let mut buffer = BytesMut::with_capacity(32);
-
-        buffer.put(&b"26 abcdefghijklmnopqrstuvwxyzand here we are"[..]);
-        let result = decoder.decode(&mut buffer);
-        assert_eq!(Ok(None), result.map_err(|_| false));
-        let result = decoder.decode(&mut buffer);
-
-        assert!(result.is_err());
-        assert_eq!(b"and here we are"[..], buffer);
-    }
-
-    #[test]
-    fn octet_decode_rejects_exceeded_frame_length_multiple_frames() {
-        let mut decoder = SyslogDecoder::new(16);
-        let mut buffer = BytesMut::with_capacity(32);
-
-        buffer.put(&b"26 abc"[..]);
-        let _result = decoder.decode(&mut buffer);
-
-        buffer.put(&b"defghijklmnopqrstuvwxyzand here we are"[..]);
-        let result = decoder.decode(&mut buffer);
-
-        println!("{:?}", result);
-        assert!(result.is_err());
-        assert_eq!(b"and here we are"[..], buffer);
-    }
-
-    #[test]
-    fn octet_decode_moves_past_exceeded_frame_length_multiple_frames() {
-        let mut decoder = SyslogDecoder::new(16);
-        let mut buffer = BytesMut::with_capacity(32);
-
-        buffer.put(&b"32thisshouldbelongerthanthmaxframeasizewhichmeansthesyslogparserwillnotbeabletodecodeit"[..]);
-        let _ = decoder.decode(&mut buffer);
-
-        assert_eq!(decoder.octet_decoding, Some(State::DiscardingToEol));
-        buffer.put(&b"wemustcontinuetodiscard\n32 something valid"[..]);
-        let result = decoder.decode(&mut buffer);
-
-        assert!(result.is_err());
-        assert_eq!(b"32 something valid"[..], buffer);
     }
 }
